ai-triage-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
google-generativeai==0.4.1
python-dotenv==1.0.1
